+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.3.1.0  -- 2024-06-24
 
+* Expose SessionT, lift all functions to MonadIO
 * Add Session.getAllVersionedDocs
 
 ## 0.3.0.0  -- 2024-04-04